fix: Use whole seconds for created_at timestamp in replication events
Changed the created_at timestamp in ReplicationEventResourceContext from
float (with decimal precision) to int (whole seconds). This removes
unnecessary decimal precision from the Unix timestamp which was causing
values like 1764325181.1376667 to be stored.

The latency calculation in kafka_consumer.py will still work correctly
since it subtracts the integer timestamp from time.time(), giving
sub-second precision in the resulting latency metric.

Updated tests to use unittest.mock.ANY for created_at field matching
since the timestamp is dynamically generated.

+++ rbac/management/relation_replicator/relation_replicator.py
@@ -139,7 +139,7 @@ class ReplicationEventResourceContext:
     resource_id: str | None
     org_id: str
     event_type: str
-    created_at: float  # Unix timestamp when event was created
+    created_at: int  # Unix timestamp when event was created (whole seconds)
 
     def __init__(
         self,
@@ -147,7 +147,7 @@ class ReplicationEventResourceContext:
         event_type: str,
         resource_type: str | None = None,
         resource_id: str | None = None,
-        created_at: float | None = None,
+        created_at: int | None = None,
     ):
         """Initialize ReplicationEventResourceContext."""
         self.resource_type = resource_type
@@ -155,7 +155,7 @@ class ReplicationEventResourceContext:
         self.org_id = org_id
         self.event_type = event_type
         # Capture creation timestamp for latency tracking
-        self.created_at = created_at if created_at is not None else time.time()
+        self.created_at = created_at if created_at is not None else int(time.time())
 
     def to_json(self) -> Dict[str, object]:
         """Convert to JSON dictionary."""
